Abstract Class Refactor
Refactored interfaces to abstract classes.

Completed for:
Identity
AzureActiveDirectory
Okta
GoogleWorkspace

Refactored namespace specific PropertyNameAttribute classes into DirectorySystemPropertyNameAttribute class.

ActiveDirectory namespace not complete!

+++ Galactic.Identity.AzureActiveDirectory/Group.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Reflection;
-using System.Linq;
-using System.Collections;
-using System.Collections.Generic;
-using GraphGroup = Microsoft.Graph.Group;
+﻿using GraphGroup = Microsoft.Graph.Group;
 
 namespace Galactic.Identity.AzureActiveDirectory
 {
-    public class Group : IGroup
+    public class Group : Identity.Group
     {
         // ----- CONSTANTS -----
 
@@ -22,7 +17,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// All users that are a member of this group or a subgroup.
         /// </summary>
-        public List<IUser> AllUserMembers
+        public override List<Identity.User> AllUserMembers
         {
             get
             {
@@ -33,7 +28,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// Groups that are a member of the group.
         /// </summary>
-        public List<IGroup> GroupMembers
+        public override List<Identity.Group> GroupMembers
         {
             get
             {
@@ -44,15 +39,15 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// The members of the group.
         /// </summary>
-        public List<IIdentityObject> Members
+        public override List<IdentityObject> Members
         {
             get
             {
                 //Members.AddRange(UserMembers.ConvertAll<IIdentityObject>(member => member));
                 //Members.AddRange(GroupMembers.ConvertAll<IIdentityObject>(member => member));
-                List <IIdentityObject> temp = new ();
-                temp.AddRange(UserMembers.ConvertAll<IIdentityObject>(member => member));
-                temp.AddRange(GroupMembers.ConvertAll<IIdentityObject>(member => member));
+                List <IdentityObject> temp = new ();
+                temp.AddRange(UserMembers);
+                temp.AddRange(GroupMembers);
 
 
                 return temp;
@@ -60,20 +55,9 @@ namespace Galactic.Identity.AzureActiveDirectory
         }
 
         /// <summary>
-        /// The number of members in the group.
-        /// </summary>
-        public int MemberCount
-        {
-            get
-            {
-                return Members.Count;
-            }
-        }
-
-        /// <summary>
         /// Users that are a member of the group. (Not including subgroups.)
         /// </summary>
-        public List<IUser> UserMembers
+        public override List<Identity.User> UserMembers
         {
             get
             {
@@ -84,8 +68,8 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// The date and time that the object was created.
         /// </summary>
-        [GraphPropertyName("createdDateTime")]
-        public DateTime? CreationTime
+        [DirectorySystemPropertyName("createdDateTime")]
+        public override DateTime? CreationTime
         {
             get
             {
@@ -96,7 +80,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// The list of groups this object is a member of.
         /// </summary>
-        public List<IGroup> Groups
+        public override List<Identity.Group> Groups
         {
             get
             {
@@ -107,7 +91,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// The type or category of the object. Empty if unknown.
         /// </summary>
-        public string Type
+        public override string Type
         {
             get
             {
@@ -118,8 +102,8 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// The object's unique ID in the system.
         /// </summary>
-        [GraphPropertyName("id")]
-        public string UniqueId
+        [DirectorySystemPropertyName("id")]
+        public override string UniqueId
         {
             get
             {
@@ -130,8 +114,8 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// A description of the object.
         /// </summary>
-        [GraphPropertyName("description")]
-        public string Description
+        [DirectorySystemPropertyName("description")]
+        public override string Description
         {
             get
             {
@@ -151,7 +135,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// The name of the group.
         /// </summary>
-        [GraphPropertyName("displayName")]
+        [DirectorySystemPropertyName("displayName")]
         public string DisplayName
         {
             get
@@ -172,7 +156,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// Primary email alias of group.
         /// </summary>
-        [GraphPropertyName("mail")]
+        [DirectorySystemPropertyName("mail")]
         public string PrimaryEmailAddress
         {
             get
@@ -188,7 +172,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// All proxy addresses on the group.
         /// </summary>
-        [GraphPropertyName("proxyAddresses")]
+        [DirectorySystemPropertyName("proxyAddresses")]
         public List<string> EmailAddresses
         {
             get
@@ -204,7 +188,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// True if group is mail enabled, otherwise false.
         /// </summary>
-        [GraphPropertyName("mailEnabled")]
+        [DirectorySystemPropertyName("mailEnabled")]
         public bool MailEnabled
         {
             get
@@ -216,7 +200,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// The mail nickname of the group.
         /// </summary>
-        [GraphPropertyName("mailNickname")]
+        [DirectorySystemPropertyName("mailNickname")]
         public string MailNickname
         {
             get
@@ -237,7 +221,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// True if group is security enabled, otherwise false.
         /// </summary>
-        [GraphPropertyName("securityEnabled")]
+        [DirectorySystemPropertyName("securityEnabled")]
         public bool SecurityEnabled
         {
             get
@@ -249,7 +233,7 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// <summary>
         /// Visibility of the group.
         /// </summary>
-        [GraphPropertyName("visibility")]
+        [DirectorySystemPropertyName("visibility")]
         public string Visability
         {
             get
@@ -298,15 +282,19 @@ namespace Galactic.Identity.AzureActiveDirectory
         /// </summary>
         /// <param name="members">The members to add.</param>
         /// <returns>True if the members were added, false otherwise.</returns>
-        public bool AddMembers(List<IIdentityObject> members)
+        public override bool AddMembers(List<IdentityObject> members)
         {
             if(members != null)
             {
-                foreach(IIdentityObject member in members)
+                foreach(IdentityObject member in members)
                 {
-                    if(!aad.AddObjectToGroup(member.UniqueId, UniqueId))
+                    // Skip non-AAD IdentityObjects.
+                    if (member is Group || member is User)
                     {
-                        return false;
+                        if (!aad.AddObjectToGroup(member.UniqueId, UniqueId))
+                        {
+                            return false;
+                        }
                     }
                 }
                 return true;
@@ -318,65 +306,11 @@ namespace Galactic.Identity.AzureActiveDirectory
         }
 
         /// <summary>
-        /// Clears all members from this group.
-        /// </summary>
-        /// <returns>True if all members were cleared, false otherwise.</returns>
-        public bool ClearMembership()
-        {
-            foreach(IIdentityObject obj in Members)
-            {
-                if(!aad.DeleteObjectFromGroup(obj.UniqueId, UniqueId))
-                {
-                    return false;
-                }
-            }
-
-            return true;
-        }
-
-        /// <summary>
-        /// Gets the values of the attributes associated with the supplied names.
-        /// </summary>
-        /// <param name="names">The names of the attributes to get the values of.</param>
-        /// <returns>A list of identity attributes that contain the attribute's name and value, or null if no values could be returned.</returns>
-        public List<IdentityAttribute<Object>> GetAttributes(List<string> names)
-        {
-            // Create a list of IdentityAttributes to return.
-            List<IdentityAttribute<object>> attributes = new();
-
-            if (names != null)
-            {
-                // Create a dictionary of properties in this class keyed by name.
-                PropertyInfo[] propertyInfoList = typeof(User).GetProperties();
-                Dictionary<string, PropertyInfo> properties = new();
-                foreach (PropertyInfo propertyInfo in propertyInfoList)
-                {
-                    foreach (GraphPropertyNameAttribute attribute in propertyInfo.GetCustomAttributes<GraphPropertyNameAttribute>())
-                    {
-                        properties.Add(attribute.Name, propertyInfo);
-                    }
-                }
-
-                // Fill the list of IdentityAttributes with the name and value of the attribute with the supplied name.
-                foreach (string name in names)
-                {
-                    if (properties.ContainsKey(name))
-                    {
-                        attributes.Add(new(name, properties[name].GetValue(this)));
-                    }
-                }
-            }
-
-            // Return the attributes found.
-            return attributes;
-        }
-
-        /// <summary>
         /// Sets attribute values of an identity object. If null or empty values are supplied the attribute's value will be deleted.
         /// </summary>
         /// <param name="attributes">The attribute to set.</param>
         /// <returns>A list of identity attributes that have values of true if the attribute was set successfully, or false otherwise.</returns>
-        public List<IdentityAttribute<bool>> SetAttributes(List<IdentityAttribute<Object>> attributes)
+        public override List<IdentityAttribute<bool>> SetAttributes(List<IdentityAttribute<Object>> attributes)
         {
             List<IdentityAttribute<bool>> results = new List<IdentityAttribute<bool>>();
 
@@ -389,49 +323,23 @@ namespace Galactic.Identity.AzureActiveDirectory
         }
 
         /// <summary>
-        /// Returns an enumerator that iterates through the collection.
-        /// </summary>
-        /// <returns>An IEnumerator object that can be used to iterate through the collection.</returns>
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-
-        /// <summary>
-        /// Returns an enumerator that iterates through the collection.
-        /// </summary>
-        /// <returns>An IEnumerator object that can be used to iterate through the collection.</returns>
-        IEnumerator<IIdentityObject> IEnumerable<IIdentityObject>.GetEnumerator()
-        {
-            return ((IGroup)this).GetEnumerator();
-        }
-
-        /// <summary>
-        /// Returns an enumerator that iterates through the collection.
-        /// </summary>
-        /// <returns>An IEnumerator object that can be used to iterate through the collection.</returns>
-        public IEnumerator<IIdentityObject> GetEnumerator()
-        {
-            foreach (IIdentityObject member in Members)
-            {
-                yield return member;
-            }
-        }
-
-        /// <summary>
         /// Removes identity objects from the group.
         /// </summary>
         /// <param name="members">The objects to remove.</param>
         /// <returns>True if the objects were removed, false otherwise.</returns>
-        public bool RemoveMembers(List<IIdentityObject> members)
+        public override bool RemoveMembers(List<IdentityObject> members)
         {
             if (members != null)
             {
-                foreach (IIdentityObject member in members)
+                foreach (IdentityObject member in members)
                 {
-                    if (!aad.DeleteObjectFromGroup(member.UniqueId, UniqueId))
+                    // Skip non-AAD members.
+                    if (member is Group || members is User)
                     {
-                        return false;
+                        if (!aad.DeleteObjectFromGroup(member.UniqueId, UniqueId))
+                        {
+                            return false;
+                        }
                     }
                 }
                 return true;
@@ -441,94 +349,6 @@ namespace Galactic.Identity.AzureActiveDirectory
                 throw new ArgumentNullException();
             }
         }
-
-        /// <summary>
-        /// Compares this identity object to another identity object.
-        /// </summary>
-        /// <param name="other">The other identity object to compare this one to.</param>
-        /// <returns>1 iif the object supplied comes before this one in the sort order, 0 if they occur at the same position, 1 if the object supplied comes after this one in the sort order.</returns>
-        public int CompareTo(IIdentityObject other)
-        {
-            return ((IIdentityObject)this).CompareTo(other);
-        }
-
-        /// <summary>
-        /// Compares this User to another User.
-        /// </summary>
-        /// <param name="other">The other User to compare this one to.</param>
-        /// <returns>-1 if the object supplied comes before this one in the sort order, 0 if they occur at the same position, 1 if the object supplied comes after this one in the sort order</returns>
-        public int CompareTo(Group other)
-        {
-            if (other != null)
-            {
-                return string.Compare(UniqueId.ToString(), other.UniqueId.ToString(), StringComparison.OrdinalIgnoreCase);
-            }
-            else
-            {
-                throw new ArgumentNullException("other");
-            }
-        }
-
-        /// <summary>
-        /// Checks whether x and y are equal (have the same UniqueIds).
-        /// </summary>
-        /// <param name="x">The first identity object to check.</param>
-        /// <param name="y">The second identity object to check.</param>
-        /// <returns>True if the identity objects are equal, false otherwise.</returns>
-        public bool Equals(IIdentityObject x, IIdentityObject y)
-        {
-            return ((IIdentityObject)this).Equals(x, y);
-        }
-
-        /// <summary>
-        /// Checks whether x and y are equal (using GUIDs).
-        /// </summary>
-        /// <param name="x">The first User to check.</param>
-        /// <param name="y">The second User to check against.</param>
-        /// <returns>True if the objects are equal, false otherwise.</returns>
-        public bool Equals(Group x, Group y)
-        {
-            if (x != null && y != null)
-            {
-                return x.UniqueId.Equals(y.UniqueId);
-            }
-            else
-            {
-                if (x == null)
-                {
-                    throw new ArgumentNullException("x");
-                }
-                else
-                {
-                    throw new ArgumentNullException("y");
-                }
-            }
-        }
-
-        /// <summary>
-        /// Generates a hash code for the identity object supplied.
-        /// </summary>
-        /// <param name="obj">The identity object to generate a hash code for.</param>
-        /// <returns>An integer hash code for the identity object.</returns>
-        public int GetHashCode(IIdentityObject obj) => IIdentityObject.GetHashCode(obj);
-
-        /// <summary>
-        /// Generates a hash code for the User supplied.
-        /// </summary>
-        /// <param name="obj">The User to generate a hash code for.</param>
-        /// <returns>An integer hash code for the object.</returns>
-        public int GetHashCode(Group obj)
-        {
-            if (obj != null)
-            {
-                return obj.UniqueId.GetHashCode();
-            }
-            else
-            {
-                throw new ArgumentNullException("obj");
-            }
-        }
-
     }
 }
